backend/requirements.txt
# Flask Core
Flask==2.3.3
Flask-Cors==6.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.1.0
Flask-Limiter==3.9.2
Werkzeug==3.1.4

# Database
SQLAlchemy==2.0.44

# LandingAI and parsing
landingai-ade==0.21.2
pydantic==2.12.4
pydantic_core==2.41.5
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0

# Testing
pytest==8.4.1

# Dependencies (auto-installed)
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
